% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupstyle.R
\name{build_line_break_manipulators}
\alias{build_line_break_manipulators}
\title{Build a list of functions to address line breaks}
\usage{
build_line_break_manipulators(scope, strict)
}
\arguments{
\item{scope}{Ordered factor scalar, see \code{\link{commastyler_style}}}

\item{strict}{Logical scalar, see \code{\link{commastyler_style}}}
}
\value{
List of functions or NULL
importFrom tibble lst
}
\description{
Build a list of functions to address line breaks
}
