% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setupstyle.R
\name{commastyler_reindention}
\alias{commastyler_reindention}
\title{Define reindention input for \code{commastyler_style}}
\usage{
commastyler_reindention()
}
\value{
List of three elements:
  \describe{
    \item{\code{regex_pattern}}{NULL}
    \item{\code{indention}}{default 0}
    \item{\code{comments_only}}{TRUE}
  }
}
\description{
Define reindention input for \code{commastyler_style}
}
